requirements.txt
streamlit
python-pptx
pandas
numpy
